Update README.md
Averaging Grades

Write a program that will read in 3 grades from the keyboard and will print the average (to 2 decimal places) of those grades to the screen. It should include good prompts and labeled output. Use the examples from the earlier labs to help you. You will want to begin with a design. Notice in the sample run that the answer is stored in fixed point notation with two decimal points of precision.

Sample run:

Please input the first grade
97
Please input the second grade
98.3
Please input the third grade
95
The average of the three grades is
96.77